code/agents/03-payments-support/requirements.txt
google-adk>=1.18.0,<2.0.0
google-genai>=1.21.1,<2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
websockets>=13.0
